dle: Restructure to support multiple backends
