+++ devito/dle/backends/__init__.py
@@ -1,0 +1,5 @@
+from devito.dle.backends.common import *  # noqa
+from devito.dle.backends.utils import *  # noqa
+from devito.dle.backends.basic import BasicRewriter  # noqa
+from devito.dle.backends.advanced import DevitoRewriter, DevitoSpeculativeRewriter  # noqa
+from devito.dle.backends.yask import YaskRewriter  # noqa
