+++ devito/dle/backends/yask.py
@@ -1,0 +1,7 @@
+from devito.dle.backends import BasicRewriter, dle_pass
+
+
+class YaskRewriter(BasicRewriter):
+
+    def _pipeline(self, state):
+        super(YaskRewriter, self)._pipeline(state)
